Create task_2
In this task, start by opening the text file using the import_file variable, the with keyword, and the open() function with the "r" parameter. Be sure to replace the ### YOUR CODE HERE ### with your own code.

For now, you'll write the first line of the with statement. Running this code will produce an error because it will only contain the first line of the with statement; you'll complete this with statement in the task after this.