Create task_4
After reading the file, reassign the ip_addresses variable so its data type is updated from a string to a list. Use the .split() method to achieve this. Adding this step will allow you to iterate through each of the IP addresses in the allow list instead of navigating a large string that contains all the addresses merged together.

Afterwards, display the ip_addresses variable to verify that the update took place.