Create task_8
In this task, you'll verify that the original file was rewritten using the correct list.

Write another with statement, this time to read in the updated file. Start by opening the file. Then read the file and store its contents in the text variable.

Afterwards, display the text variable to examine the result.